requirements.txt
pandas==2.2.3
Pillow==11.1.0
plotly==5.24.1
python-dotenv==1.0.1
python_dateutil==2.9.0
Requests==2.32.3
streamlit==1.41.1
urllib3==2.3.0
